fix: auth actions handled correctly
