+++ apps/admin/src/app/(app)/(site)/layout.tsx
@@ -1,7 +1,7 @@
 import { redirect } from "next/navigation";
 import NavbarAvatar from "@/components/nav/navbar-avatar";
 import ThemeSwitcher from "@/components/nav/theme-switcher";
-import { getSession } from "@/lib/auth/auth";
+import { getSession } from "@/lib/auth/auth.server";
 import { AreaChartIcon, CpuIcon, HomeIcon, Users2Icon } from "lucide-react";
 
 import { Navbar, Sidebar } from "@energyleaf/ui/components/nav";
