+++ apps/web/src/actions/peak.ts
@@ -6,7 +6,7 @@ import type { z } from "zod";
 import "server-only";
 
 import { cookies } from "next/headers";
-import { getSession } from "@/lib/auth/auth";
+import { getActionSession } from "@/lib/auth/auth.action";
 import { addOrUpdatePeakCookieStore, isDemoUser } from "@/lib/demo/demo";
 import type { peakSchema } from "@/lib/schema/peak";
 
@@ -14,7 +14,7 @@ import { addOrUpdatePeak as addOrUpdatePeakDb } from "@energyleaf/db/query";
 import { UserNotLoggedInError } from "@energyleaf/lib/errors/auth";
 
 export async function addOrUpdatePeak(data: z.infer<typeof peakSchema>, sensorId: string, timestamp: string) {
-    const { session } = await getSession();
+    const { session } = await getActionSession();
 
     if (!session) {
         throw new UserNotLoggedInError();
