+++ apps/web/src/components/recommendations/avg-energy-consumption-comparison.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
-import { getSession } from "@/lib/auth/auth";
+import { getSession } from "@/lib/auth/auth.server";
 import {
     getAvgEnergyConsumptionForSensor,
     getAvgEnergyConsumptionForUserInComparison,
