+++ apps/web/src/lib/auth/auth.config.ts
@@ -1,0 +1,26 @@
+import { env } from "@/env.mjs";
+import { Lucia } from "lucia";
+
+import { adapter } from "@energyleaf/db/adapter";
+
+export const lucia = new Lucia(adapter, {
+    sessionCookie: {
+        name: "auth_session",
+        expires: false,
+        attributes: {
+            domain: `.${env.VERCEL_URL || env.NEXTAUTH_URL || "localhost"}`,
+            sameSite: "lax",
+            secure: env.VERCEL_ENV === "production" || env.VERCEL_ENV === "preview",
+        },
+    },
+    getUserAttributes: (attributes) => {
+        return {
+            id: attributes.id,
+            name: attributes.name,
+            email: attributes.email,
+            created: attributes.created,
+            isAdmin: attributes.isAdmin,
+            isActive: attributes.isActive,
+        };
+    },
+});
