+++ apps/web/src/lib/auth/auth.server.ts
@@ -1,0 +1,7 @@
+import { cache } from "react";
+
+import { getActionSession } from "./auth.action";
+
+import "server-only";
+
+export const getSession = cache(async () => getActionSession());
